requirements.txt
# Deep Learning Frameworks
tensorflow>=2.8.0
torch>=1.9.0
torchvision>=0.10.0
keras>=2.8.0

# Scientific Computing
numpy>=1.21.0
scipy>=1.7.0
scikit-learn>=0.24.0

# Image Processing
opencv-python>=4.5.0
Pillow>=8.2.0

# Natural Language Processing
nltk>=3.6.0
rake-nltk>=1.0.0

# Visualization
matplotlib>=3.4.0
seaborn>=0.11.0

# Environment and Utilities
python-dotenv>=0.19.0
tqdm>=4.62.0

# Optional but recommended
ipython>=7.16.0
scikit-learn
imblearn
 
uvicorn 
torch 
torchvision 
Pillow
python-multipart
fastapi>=0.68.0
transformers